+++ test/unit/extend.test.js
@@ -1,13 +1,13 @@
-import extendsFn from '../../src/js/extends.js';
+import extendFn from '../../src/js/extend.js';
 
-q.module('extends.js');
+q.module('extend.js');
 
 test('should add implicit parent constructor call', function(){
   var superCalled = false;
   var Parent = function() {
     superCalled = true;
   };
-  var Child = extendsFn(Parent, {
+  var Child = extendFn(Parent, {
       foo: 'bar'
   });
   var child = new Child();