+++ CHANGELOG.md
@@ -138,6 +138,7 @@ CHANGELOG
 * @heff Fixed double loadstart and ready events ([view](https://github.com/videojs/video.js/pull/2605))
 * @gkatsev fixed potential double default style elements ([view](https://github.com/videojs/video.js/pull/2619))
 * @imbcmdth extended createTimeRange to support multiple timeranges ([view](https://github.com/videojs/video.js/pull/2604))
+* @misteroneill rename &quot;extends&quot; to &quot;extend&quot; for ie8 ([view](https://github.com/videojs/video.js/pull/2624))
 
 --------------------